refactor: ResultMessage 패키지 변경
- ResultMessage는 DTO보다는 VO로 보는게 맞다고 판단.
